Visualdepth should take glue stones into account
Whenever there is a variation, for example 2 continuations from move A
- A’ and A’’.  When visually drawn then A’ will be the main line and
A’’ will be a child variation. When drawing visually the A’’ we will
show a line for visual aid, connection A with A’’. We need to ensure
that we have space for that line, I call that space a Glue Stone as it
takes up a stone amount of space.