Don't create a dummy node while parsing.
When parsing the file a dummy node would get created if it contained game properties but no node properties. Later on this resulted in having 1 extra dummy node in the beginning of the game. Now instead lets not create that. This can result in having no nodes in the game and lots of null checks are needed. To get around that I create a dummy node in case of a file that has no nodes.